+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Recent and upcoming changes to lightdash
 
 ## Unreleased
+### Added
+- Users can create and revoke invite links
+
 ## [0.6.3] - 2021-08-10
 ### Fixed
 - Fixed bugs with filters when selecting multiple filters on the same column